Split mocking and read_only from mocking
